+++ js/mobile.js
@@ -262,27 +262,17 @@ function createBooking() {
 	var destinationAirport = document.getElementById("destinationAirport").value  ;
 	var via1Airport = document.getElementById("via1Airport").value  ;
 	var via2Airport = document.getElementById("via2Airport").value  ;
+	// Display the spinner if any
+	var spinnerElement = document.getElementById('spinner') ;
+	if (spinnerElement) spinnerElement.classList.remove('d-none');
 
 	var XHR=new XMLHttpRequest();
-
-if (false) {
-	// Check whether a solo flight is allowed on this plane
-	if (!userIsInstructor && instructorId <= 0) {
-		for (var i = 0; i < allPlanes.length; i++)
-			if (allPlanes[i].id == plane) {
-				if (!allPlanes[i].qualifications_requises)
-					if (!confirm("Au vu de votre carnet de vol et de vos validités/annotations club, vous n'avez pas la possibilité de réserver cet avion.\nVoulez-vous malgré tout réserver?\n\n" +
-						"OK pour continuer (les instructeurs seront prévenus), Cancel pour ne pas réserver")) {
-						hideEditBookingDetails() ;
-						return ; 
-						}
-			}
-	}
-	
-}
 	XHR.onreadystatechange = function() {
 		if(this.readyState  == 4) {
 			if(this.status  == 200) {
+				// Hide the spinner if any
+				var spinnerElement = document.getElementById('spinner') ;
+				if (spinnerElement) spinnerElement.classList.add('d-none');
 				try {
 					var response = eval('(' + this.responseText.trim() + ')') ;
 				} catch(err) {