+++ mobile_header5.php
@@ -358,8 +358,8 @@ if ($userIsFlightManager && $userId == 62) {
 <?php
 if ($userId > 0) {
 ?>
-            <li><a class="dropdown-item" href="mobile_reservation.php"><i class="bi bi-display"></i> Toutes les réservations (plein écran)</a></i></li>
-            <li><a class="dropdown-item" href="mobile_today.php">&#128241; Toutes les réservations (smartphone)</a></li>
+            <li><a class="dropdown-item" href="mobile_reservation.php"><i class="bi bi-display"></i> Toutes les réservations (calendrier)</a></i></li>
+            <li><a class="dropdown-item" href="mobile_today.php">&#128241; Toutes les réservations par heure (smartphone)</a></li>
             <li><a class="dropdown-item" href="mobile_resa_by_plane.php"><i class="bi bi-airplane"></i> Toutes les réservations par avion (smartphone)</a></li>
             <li><a class="dropdown-item" href="mobile.php"><i class="bi bi-file-person-fill"></i> Mes réservations</a></li>
             <li><a class="dropdown-item" href="mobile_book.php"><i class="bi bi-plus-square-fill"></i> Nouvelle réservation</a></li>